fix: changes to copy tests folder. checkNodeVersion to src.
